refactored onboarding
