+++ src/components/Onboarding/Form/PlantCalibrateBtn.tsx
@@ -1,0 +1,14 @@
+
+export const PlantCalibrateBtn = () => {
+
+    return (
+        <>
+            <button
+                type="submit"
+                className="w-full px-5 py-3 text-base font-medium text-center text-white bg-primary-700 rounded-lg hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 sm:w-auto dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800"
+            >
+                Potrdi
+            </button>
+        </>
+    )
+}
